requirements.txt
python-dotenv
requests
groq
supabase
twilio
google-generativeai
fastapi
uvicorn
pydantic
reportlab
schedule
psycopg2-binary

